ai-agentless-recipe-shoplist/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-recipe-shoplist"
version = "0.1.0"
description = "AI-powered recipe ingredient crawler with grocery store price optimization (Python 3.11+)"
authors = [{name = "Fernando Ibanez", email = "fernando@example.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.119.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.2",
    "pydantic>=2.4.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "reportlab>=4.0.4",
    "openai>=1.3.0",
    "ollama>=0.1.7",
    "requests>=2.31.0",
    "anthropic>=0.25.0",
    "tiktoken>=0.12.0",
    "azure-identity>=1.14.0",
    "azure-ai-inference>=1.0.0b3",
    "tenacity>=8.2.0",
    "rich>=14.2.0",
    "cachetools>=6.2.1",
    "selenium>=4.15.0",
    "playwright>=1.40.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.24.4",
    "pandas>=2.0.3",
    "joblib>=1.5.2",
    "pillow>=10.0.1",
    "PyGithub>=1.59.0",
    "jmespath>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.9.1",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "black>=23.9.1",
    "isort>=5.12.0", 
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]